+++ packages/example-app-browser-ts-assets/src/typings.d.ts
@@ -1,0 +1,3 @@
+declare module 'example-lib-universal-js';
+declare module 'example-lib-universal-ts';
+declare const module;